*** empty log message ***

+++ src/print.c
@@ -56,7 +56,7 @@ void pfileinfo(char **argv)
     if (!(stat(fn, &sb))) {
       timeinfo = localtime(&sb.st_mtime);
       printf(";%13zu  %02d:%02d.%02d %02d-%02d-%02d %s\n",
-	     sb.st_size, timeinfo->tm_hour, timeinfo->tm_min,
+	     (size_t) sb.st_size, timeinfo->tm_hour, timeinfo->tm_min,
              timeinfo->tm_sec,
              timeinfo->tm_year+1900, timeinfo->tm_mon+1, timeinfo->tm_mday, fn);
     }
